+++ src/main/kotlin/csense/idea/base/bll/psiWrapper/class/operations/any/AnySuperClassType.kt
@@ -1,0 +1,15 @@
+package csense.idea.base.bll.psiWrapper.`class`.operations.any
+
+import csense.idea.base.bll.psiWrapper.`class`.*
+import csense.idea.base.bll.psiWrapper.`class`.operations.*
+
+inline fun KtPsiClass.anySuperClassType(
+    predicate: (KtPsiClass) -> Boolean
+): Boolean {
+    forEachSuperClassType { it: KtPsiClass ->
+        if (predicate(it)) {
+            return@anySuperClassType true
+        }
+    }
+    return false
+}